mirror .venv into _vendor
