+++ packages/python/src/install.ts
@@ -1,9 +1,9 @@
 import execa from 'execa';
 import fs from 'fs';
 import os from 'os';
-import { join } from 'path';
+import { dirname, join } from 'path';
 import which from 'which';
-import { Meta, debug } from '@vercel/build-utils';
+import { FileFsRef, Files, Meta, debug, glob } from '@vercel/build-utils';
 import { getVenvPythonBin, getVenvPipBin, createVenvEnv } from './utils';
 
 const isWin = process.platform === 'win32';
@@ -373,6 +373,7 @@ async function pipInstall(
       return;
     } catch (err) {
       console.log(`Failed to run "${prettyUv}", falling back to pip`);
+      console.log(err instanceof Error ? err.message : String(err)); // TODO remove
       debug(`error: ${err}`);
     }
   }
@@ -625,4 +626,61 @@ export async function exportRequirementsFromPipfile({
   await fs.promises.writeFile(outPath, stdout);
   debug(`Exported pipfile requirements to ${outPath}`);
   return outPath;
+}
+
+export async function mirrorSitePackagesIntoVendor({
+  venvPath,
+  vendorBaseDir,
+  vendorDirName,
+}: {
+  venvPath: string;
+  vendorBaseDir: string;
+  vendorDirName: string;
+}): Promise<Files> {
+  const vendorFiles: Files = {};
+  const vendorDirOnDisk = join(vendorBaseDir, vendorDirName);
+
+  // Mirror the venv's site-packages into a vendored directory under
+  // `.vercel/python/.../_vendor`
+  // this is to avoid bundling the entire virtual environment into Lambda.
+  try {
+    try {
+      await fs.promises.rm(vendorDirOnDisk, { recursive: true, force: true });
+    } catch (err: any) {
+      debug('Failed to remove existing vendor directory', err);
+    }
+
+    await fs.promises.mkdir(vendorDirOnDisk, { recursive: true });
+
+    const sitePackageDirs = await getVenvSitePackagesDirs(venvPath);
+    for (const dir of sitePackageDirs) {
+      if (!fs.existsSync(dir)) continue;
+
+      const dirFiles = await glob('**', dir);
+      for (const relativePath of Object.keys(dirFiles)) {
+        if (
+          relativePath.endsWith('.pyc') ||
+          relativePath.includes('__pycache__')
+        ) {
+          continue;
+        }
+
+        const srcFsPath = join(dir, relativePath);
+        const destFsPath = join(vendorDirOnDisk, relativePath);
+        await fs.promises.mkdir(dirname(destFsPath), { recursive: true });
+        await fs.promises.copyFile(srcFsPath, destFsPath);
+
+        const bundlePath = join(vendorDirName, relativePath).replace(
+          /\\/g,
+          '/'
+        );
+        vendorFiles[bundlePath] = new FileFsRef({ fsPath: destFsPath });
+      }
+    }
+  } catch (err) {
+    console.log('Failed to collect site-packages from virtual environment');
+    throw err;
+  }
+
+  return vendorFiles;
 }
